requirements.txt
asgiref==3.2.7
astroid==2.3.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.2
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
crypto==1.4.1
cryptography==2.9
decorator==4.4.2
Django==3.0.6
docxtpl==0.9.2
future==0.18.2
huepy==1.2.1
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
instabot==0.117.0
isort==4.3.21
Jinja2==2.11.1
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.4.0
moviepy==1.0.3
Naked==0.1.31
numpy==1.18.2
packaging==20.4
pandas==1.0.3
Pillow==7.1.1
pluggy==0.13.1
proglog==0.1.9
py==1.8.2
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.7
pylint==2.4.4
pyparsing==2.4.7
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.13.0
python-telegram-bot==12.6.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.9.1
responses==0.10.15
rsa==4.6
schedule==0.6.0
shellescape==3.8.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tornado==6.0.4
tqdm==4.46.1
urllib3==1.25.8
urwid==2.1.0
wcwidth==0.2.4
wrapt==1.11.2
xlrd==1.2.0
